dashboard-app/requirements.txt
streamlit
pandas
numpy
statsmodels
matplotlib
seaborn
requests
python-dotenv
